fix: Encoding of opaque subgraphs with no associated qubit/bit (#1295)
Fixes #1294.

This PR addresses a problem when emitting subgraphs that cannot be put
in a pytket barrier since they were not associated with registers.

Given the following graph:
```mermaid
graph LR
    subgraph 0 ["(0) Module"]
        direction LR
        subgraph 1 ["(1) [**FuncDefn: #quot;test#quot;**]"]
            direction LR
            style 1 stroke:#832561,stroke-width:3px
            2["(2) Input"]
            3["(3) Output"]
            16["(16) Unsupported OP"]
            17["(17) tket.rotation.from_halfturns_unchecked"]
            16--"0:0<br>float64"-->17
            17--"0:0<br>rotation"-->3
        end
    end
```
The encoding of `from_halfturns_unchecked` would request the unsupported
op to be emitted in an opaque subgraph, but we are unable to put that
operation alone in a barrier.

While #1294 proposes doing a lookahead on the wires and
backtracking/marking `from_halfturns_unchecked` as unsupported as well,
the solution here is simpler.
We just let the `EncodedCircuit` metadata carry multiple subgraphs not
linked to barriers, so we register them there.

The downside is that now the circuit above is not encodable as a
standalone pytket circuit (since the metadata stays on the rust
in-memory object), but we can change it in the future if needed.

---------

Co-authored-by: Alan Lawrence <alan.lawrence@cambridgequantum.com>