refactor: rename PytketPass to PytketHugrPass (#1310)
Renamig to make clear that the pass is a Hugr transform. For me this
helps distinguish it from the passes in the `pytket.passes` moddule.

Driveby: small improvement to `PytketHugrPass` docs.

`PytketPass` isn't in a pypi release yet hence why I'm renaming now.

Note that the change is not breaking as the package is not released yet.