docs: add notebook on optimization of Guppy programs (#1287)
closes #1284

Found a couple bugs whilst making this. Agustin has now fixed most of
them.

The examples here are super trivial at the moment but I demonstrate
`NormalizeGuppy`, `RemoveRedundancies` and `SquashRzPhasedX`.

The example using `SquashRzPhasedX` isn't quite working yet. It should
simplify to a single Rz gate but this isn't working yet due to an issue
with order edges. Should be fixed once we have the transformation in
https://github.com/Quantinuum/hugr/pull/2739 is part of
`NormalizeGuppy`.

Note I moved the `examples` directory to be part of the docs directory
so that I could include the example in the docs site more easily.