fix: Incorrect manylinux version in linux wheels (#1327)
The linux-gnu wheel builds were incorrectly stating we support
`manylinux2014` (`glibc 2.17`).
The wheels include the `tket-c-api` library, which is built directly on
the container and hence uses the system's glibc version.

For the supported `ubuntu 24.04` runner, that is `glibc 2.39`.
After some investigation, it seems that if we change the conan profile
used to build `tket-c-api` from `-gcc14` to `-gcc13` the resulting `.so`
only requires `glibc 2.34`.

This is just a patch to ensure the wheels we built correctly announce
their properties.
The issue for actually fixing the root problem is here:
https://github.com/Quantinuum/tket2/issues/1272

[Test
build](https://github.com/Quantinuum/tket2/actions/runs/20335124960/job/58419711500#step:13:314)